리스트 출력 리팩토링. (중복코드 정리. 기능별분류).  평점순 정렬 구현

+++ wifi_project/Order/src/main/java/com/wifi/order/model/Itemlist.java
@@ -25,7 +25,7 @@ public class Itemlist {
 	private int view_count;			// 조회수
 	private int category;			// 판매물품 카테고리 : 0 과일, 1 육류, 2 해산물, 3 채소, 4 생필품, 5 음료, 6 전체
 	private int midx;				// 작성자(회원번호) - FK
-	private int rvs_avg;			// 작성자 평균평점 -Item_rvs table join
+	private float rvs_avg;			// 작성자 평균평점 -Item_rvs table join
 	private int rvs_totalRow;		// 작성자 총평점개수 -Item_rvs table join
 	
 	
@@ -35,7 +35,7 @@ public class Itemlist {
 
 	public Itemlist(int iidx, String title, int price, int count_m, int count_w, Timestamp regdate, Timestamp receive,
 			int receive_check, String addr, String location, String content, String photo, int istate, int view_count,
-			int category, int midx, int rvs_avg, int rvs_totalRow) {
+			int category, int midx, float rvs_avg, int rvs_totalRow) {
 		this.iidx = iidx;
 		this.title = title;
 		this.price = price;
@@ -217,12 +217,12 @@ public class Itemlist {
 	}
 
 
-	public int getRvs_avg() {
+	public float getRvs_avg() {
 		return rvs_avg;
 	}
 
 
-	public void setRvs_avg(int rvs_avg) {
+	public void setRvs_avg(float rvs_avg) {
 		this.rvs_avg = rvs_avg;
 	}
